+++ lib/config/validation.ts
@@ -1,9 +1,4 @@
 import is from '@sindresorhus/is';
-import jsonata from 'jsonata';
-import { logger } from '../logger';
-import { allManagersList, getManagerList } from '../modules/manager';
-import { isCustomManager } from '../modules/manager/custom';
-import type { RegexManagerTemplates } from '../modules/manager/custom/regex/types';
 import { allManagersList, getManagerList } from '../modules/manager';
 import { isCustomManager } from '../modules/manager/custom';
 import type { CustomManager } from '../modules/manager/custom/types';
@@ -42,6 +37,7 @@ import * as regexOrGlobValidator from './validation-helpers/regex-glob-matchers'
 import {
   getParentName,
   isFalseGlobal,
+  validateJSONataManagerFields,
   validateNumber,
   validatePlainObject,
   validateRegexManagerFields,
@@ -833,59 +829,6 @@ export async function validateConfig(
   errors.sort(sortAll);
   warnings.sort(sortAll);
   return { errors, warnings };
-}
-
-function validateJSONataManagerFields(
-  customManager: CustomManager,
-  currentPath: string,
-  errors: ValidationMessage[],
-): void {
-  if (!is.nonEmptyString(customManager.fileFormat)) {
-    errors.push({
-      topic: 'Configuration Error',
-      message: 'Each JSONata manager must contain a fileFormat field.',
-    });
-  }
-
-  if (is.nonEmptyArray(customManager.matchStrings)) {
-    for (const matchString of customManager.matchStrings) {
-      try {
-        jsonata(matchString);
-      } catch (err) {
-        logger.debug(
-          { err },
-          'customManager.matchStrings JSONata query validation error',
-        );
-        errors.push({
-          topic: 'Configuration Error',
-          message: `Invalid JSONata query for ${currentPath}: \`${matchString}\``,
-        });
-      }
-    }
-  } else {
-    errors.push({
-      topic: 'Configuration Error',
-      message: `Each Custom Manager must contain a non-empty matchStrings array`,
-    });
-  }
-
-  const mandatoryFields = ['currentValue', 'datasource'];
-  for (const field of mandatoryFields) {
-    if (!hasField(customManager, field)) {
-      errors.push({
-        topic: 'Configuration Error',
-        message: `JSONata Managers must contain ${field}Template configuration or ${field} in the query `,
-      });
-    }
-  }
-
-  const nameFields = ['depName', 'packageName'];
-  if (!nameFields.some((field) => hasField(customManager, field))) {
-    errors.push({
-      topic: 'Configuration Error',
-      message: `JSONata Managers must contain depName or packageName in the query or their templates`,
-    });
-  }
 }
 
 /**
